+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2023-03-20
+
+### Fixed
+
+- For internal data cloning, if `structuredClone` is not available, using [devalue](https://github.com/Rich-Harris/devalue) instead.
+
 ## [0.6.0] - 2023-03-19
 
 ### Changed